Look for Proxy and StartData only in proxy.jar

+++ agent-test/src/test/java/com/github/kornilova_l/flamegraph/proxy/TestWithProxy.java
@@ -74,6 +74,9 @@ public class TestWithProxy {
     }
 
     static void clearOutputFile() {
+        if (!outputFile.exists()) {
+            return;
+        }
         if (!outputFile.delete()) {
             throw new RuntimeException("Cannot delete output file");
         }
